+++ common/src/main/java/com/gitlab/srcmc/rctmod/api/service/TrainerSpawner.java
@@ -282,11 +282,17 @@ public class TrainerSpawner {
 
     private boolean canSpawnFor(Player player) {
         var config = RCTMod.getInstance().getServerConfig();
+        var glob = config.globalSpawnChance();
+        var gmin = config.globalSpawnChanceMinimum();
+        var tcnt = this.getSpawnCount(player.getUUID());
+        var mcnt = config.maxTrainersPerPlayer();
 
-        return config.globalSpawnChance() >= player.getRandom().nextFloat()
-            && RCTMod.getInstance().getTrainerManager().getPlayerLevel(player) > 0
+        ModCommon.LOG.info(String.format("GLOBAL CHANCE: %.2f, %d/%d", gmin + (Math.max(glob, gmin) - gmin)*(1 - Math.min(1, (tcnt + 1)/(double)mcnt)), tcnt, mcnt));
+
+        return tcnt < mcnt
             && this.getSpawnCount() < config.maxTrainersTotal()
-            && this.getSpawnCount(player.getUUID()) < config.maxTrainersPerPlayer();
+            && RCTMod.getInstance().getTrainerManager().getPlayerLevel(player) > 0
+            && gmin + (Math.max(glob, gmin) - gmin)*(1 - Math.min(1, (tcnt + 1)/mcnt)) >= player.getRandom().nextFloat();
     }
 
     private void spawnFor(Player player, String trainerId, BlockPos pos) {